fix(SEO): update SEO config

+++ bienes_raices/components/last_properties/property_image.py
@@ -3,7 +3,7 @@ from bienes_raices.styles.styles import COLOR_STYLE
 
 
 def property_image(
-    image_src: str, image_alt: str, image_title: str, price: int
+    image_src: str, image_alt: str, image_title: str, price: float
 ) -> rx.Component:
     return rx.box(
         rx.image(
